Fix other competing markup reverting the quote styles
Due to the nature of codemirror decorations, the previous`.cm-quote` span had only been applying to the line up until the point that some competing formatting was applied, resulting in partially-highlighted quote lines.

As far as I can tell, short of manual dom manipulation (which might not be terrible if I could find the right callback), I can't get the `.cm-quote` class onto every element in the line as per a stock blockquote.

To get around this, I've applied the blockquote text colour to the line itself, by selecting for list lines containing quote formatting. In practice this works well, assuming stock-like expectations for quote styles.